backend/requirements.txt
fastapi
uvicorn
faiss-cpu==1.7.4
PyPDF2
numpy
ollama
python-multipart
